chore: move mentions file
