+++ packages/rocketchat-mentions/server/Mentions.js
@@ -2,7 +2,7 @@
 * Mentions is a named function that will process Mentions
 * @param {Object} message - The message object
 */
-import Mentions from '../client/Mentions';
+import Mentions from '../Mentions';
 export default class MentionsServer extends Mentions {
 	constructor(args) {
 		super(args);
